fix(diagnostic_converter): move headers to a separate directory
