+++ evaluator/diagnostic_converter/test/test_converter_node.cpp
@@ -12,8 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "converter_node.hpp"
-#include "gtest/gtest.h"
+#include "diagnostic_converter/converter_node.hpp"
+
+#include <gtest/gtest.h>
 
 #include <rclcpp/rclcpp.hpp>
